Revert "patch-file-path-sanitization"
