--- README.md
@@ -1,13 +1,0 @@
-### Changelog 
-
-#### v1.0.0
-___
-
-#### Added or Changed 
-| File | Desc | Label |
-|---------|----------|-------- |
-| [sanitization.py](./sanitization.py)  | - validate_file_path <br>  | Added | 
-| [utils.py](./utils.py)  | - extract_file_name <br>  | Added |
-| [utils.py](./utils.py)  | - save_data <br>  | Changed |
-| [history.py](./history.py)  | - file_path_sanitization <br>  | Added |
-
